fix(chrome-headless-browser-property): Refactorings. 1. splits the chrome-headless-driver tests into 2 separate classes to avoid dependency clashes. 2. makes setChromeOptions method non-static.
