+++ drone-webdriver/src/main/java/org/jboss/arquillian/drone/webdriver/factory/ChromeDriverFactory.java
@@ -114,7 +114,7 @@ public class ChromeDriverFactory extends AbstractWebDriverFactory<ChromeDriver> 
         return capabilities;
     }
 
-    public static void setChromeOptions(WebDriverConfiguration configuration, DesiredCapabilities capabilities) {
+    public void setChromeOptions(WebDriverConfiguration configuration, DesiredCapabilities capabilities) {
         ChromeOptions chromeOptions = new ChromeOptions();
 
         String browser = configuration.getBrowser().toLowerCase();
